[FIX] Fixed segmentation fault problem by checking != NULL before free()
Thanks to  Mark Nauwelaerts for patch.
Quoting Mark's email on transcode-devel:
"""
[...]
Also, as aviindex uses the AVI-library in a "rare" way, that is, it issues
AVI_open_input_file(in_file,0) (--> so, without index), this reveals that some
free()'s are being done in AVI_close() without a full check on whether these
were set in the first place.
This has been corrected as well.
[...]
"""