add the module registry.
With the introduction of the module registry, the caller/client can now
let the library automatically load the right module for a given codec/format,
without the need of specifiying the exact module.
Fallback list for a given format are allowed.

The binding between formats and modules is specified by a configuration file.