[avi][tools] better maximum size detection for avisplit
When using avisplit with the "maximum size" option on videos with large
variable gaps between keyframes many files were over the limit, some by
5Mb.

I found that the maximum size was tested against the current filesize +
1MB.  A frame size was taken and stored in the variable "bytes", but
that was just the single frame so it was very small, I imagine that plus
filesize never gave correct results so it was changed to arbitrary 1MB,
which would work on files with less than 1Mb between keyframes.

I added some code to loop through to the next keyframe (or eof) and get
the bytes between keyframes and test whether that plus filesize is over
the limit.  I didn't notice any speed difference for all the extra
loops.

Patch courtesy of `Brad' <bradley.jenkins@unix.net>