[filter][vid.stab] in crop mode, use border from previous stabilized frame.
The pixel border not suitable for stabilization are not equal to the preceding frame
as should be and is claimed in the description of the "crop" flag.
The output should be initialized using the *previous* output frame instead of the
input frame, which is not stabilized.

Pointed out by Guido Torelli.